+++ CHANGELOG.rst
@@ -2,14 +2,21 @@ This document contains the ``fslpy`` release history in reverse chronological
 order.
 
 
-2.3.0 (Under development)
--------------------------
+2.3.0 (Tuesday June 25th 2019)
+------------------------------
 
 
 Added
 ^^^^^
 
 
+* New :class:`.Bitmap` class, for loading bitmap images. The
+  :meth:`.Bitmap.asImage` method can be used to convert a ``Bitmap`` into
+  an :class:`.Image`.
+* The :class:`.Image` class now has support for the ``RGB24`` and ``RGBA32``
+  NIfTI data types.
+* New :attr:`.Image.nvals` property, for use with ``RGB24``/``RGBA32``
+  images.
 * New :meth:`.LabelAtlas.get` and :meth:`ProbabilisticAtlas.get` methods,
   which return an :class:`.Image` for a specific region.
 * The :meth:`.AtlasDescription.find` method also now a ``name`` parameter,
@@ -24,6 +31,7 @@ Fixed
 
 * The :func:`.makeWriteable` function will always create a copy of an
   ``array`` if its base is a ``bytes`` object.
+* Fixed a bug in the :meth:`.GitfitMesh.loadVertices` method.
 
 
 2.2.0 (Wednesday May 8th 2019)